Добавил метод createAd и deleteAd

+++ src/main/java/ru/skypro/homework/service/repository/AdsRepository.java
@@ -3,5 +3,8 @@ package ru.skypro.homework.service.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.skypro.homework.model.Ad;
 
-public interface AdsRepository extends JpaRepository<Ad, Long> {
+public interface AdsRepository extends JpaRepository<Ad, Integer> {
+
+    //Ad findById(int adsId);
+
 }
